Split app directory resolution
